turned HeightLimitListener.java into a task
